Fixed bug summing all zero values throwing error
"Can trust the length is equal to one..." comment was incorrect when
summing zero values, as these values get stripped out prior to where the
comment was made, but after the initial length check. This means in
cases where only zero values were summed, this assumption proves
incorrect and an error accessing the first element of an empty list
throws an error.