Improved label positioning by bypassing some calculations
The more conversions a number/unit goes through, the more room for
floating point error. Just added extra checking if the value is an
absolute unit (pixel, etc) then just use that value directly, instead
of doing a round of conversions that should bring it back to the same
value anyway (spoiler: it rarely exactly does).